minimum point on maximal path
You have a 2d matrix of integers. Find a path from the top-left (0, 0)
to bottom-right (m,n) that avoids small numbers along the way.
Specifically maximize the smallest integer on the path. You are only
allowed to move down or to the right.

Simple example (optimal path in bold). Smallest integer on path: 4

5, 1
4, 5

Complex example. Smallest integer on path:

5, 1, 6
4, 5, 8
2, 3, 9